+++ packages/yoastseo/webpack/externals/externals.js
@@ -1,0 +1,11 @@
+/**
+ * The main external dependency mapping used by the YoastSEO package.
+ *
+ * @type {Object}
+ */
+module.exports = {
+	jed: "jed",
+	"lodash-es": "lodash",
+	lodash: "lodash",
+	yoastseo: [ "yoast", "analysis" ],
+};
